requirements.txt
wandb == 0.15.2
flax == 0.7.4
optax == 0.1.7
hydra-core == 1.3.2
matplotlib == 3.7.1